refactor(account): update import paths for account components
- Changed import paths for delete-account-form, email-form, name-form, and password-form components to reflect the new directory structure.
- Centralized account-related actions and data imports under a unified path for better organization.
