+++ web-app/src/app/app/account/components/delete-account-form.tsx
@@ -6,11 +6,11 @@ import { useTranslations } from "next-intl";
 import { useForm } from "react-hook-form";
 import { toast } from "sonner";
 
-import { deleteAccount } from "@/app/(app)/dashboard/account/actions";
+import { deleteAccount } from "@/app/app/account/actions";
 import {
   DeleteAccountFormType,
   deleteAccountSchema,
-} from "@/app/(app)/dashboard/account/data";
+} from "@/app/app/account/data";
 import { Button } from "@/components/ui/button";
 import {
   Card,
