+++ web-app/src/app/app/account/components/email-form.tsx
@@ -5,11 +5,8 @@ import { useTranslations } from "next-intl";
 import { useForm } from "react-hook-form";
 import { toast } from "sonner";
 
-import { updateEmail } from "@/app/(app)/dashboard/account/actions";
-import {
-  emailFormSchema,
-  EmailFormType,
-} from "@/app/(app)/dashboard/account/data";
+import { updateEmail } from "@/app/app/account/actions";
+import { emailFormSchema, EmailFormType } from "@/app/app/account/data";
 import { Button } from "@/components/ui/button";
 import {
   Card,
